# Speed up the slow path of FastThreadLocal