--- a/microbench/src/main/java/io/netty/microbench/util/AbstractMicrobenchmark.java
+++ b/microbench/src/main/java/io/netty/microbench/util/AbstractMicrobenchmark.java
@@ -32,17 +32,6 @@ import org.openjdk.jmh.runner.options.ChainedOptionsBuilder;
 public class AbstractMicrobenchmark extends AbstractMicrobenchmarkBase {
 
     protected static final int DEFAULT_FORKS = 2;
-    protected static final String[] JVM_ARGS;
-
-    static {
-        final String[] customArgs = {
-        "-Xms768m", "-Xmx768m", "-XX:MaxDirectMemorySize=768m", "-Djmh.executor=CUSTOM",
-        "-Djmh.executor.class=io.netty.microbench.util.AbstractMicrobenchmark$HarnessExecutor" };
-
-        JVM_ARGS = new String[BASE_JVM_ARGS.length + customArgs.length];
-        System.arraycopy(BASE_JVM_ARGS, 0, JVM_ARGS, 0, BASE_JVM_ARGS.length);
-        System.arraycopy(customArgs, 0, JVM_ARGS, BASE_JVM_ARGS.length, customArgs.length);
-    }
 
     public static final class HarnessExecutor extends ThreadPoolExecutor {
         public HarnessExecutor(int maxThreads, String prefix) {
@@ -52,27 +41,36 @@ public class AbstractMicrobenchmark extends AbstractMicrobenchmarkBase {
         }
     }
 
-    private final boolean disableAssertions;
-    private String[] jvmArgsWithNoAssertions;
+    private final String[] jvmArgs;
 
     public AbstractMicrobenchmark() {
-        this(false);
+        this(false, false);
     }
 
     public AbstractMicrobenchmark(boolean disableAssertions) {
-        this.disableAssertions = disableAssertions;
+        this(disableAssertions, false);
+    }
+
+    public AbstractMicrobenchmark(boolean disableAssertions, boolean disableHarnessExecutor) {
+        final String[] customArgs;
+        if (disableHarnessExecutor) {
+            customArgs = new String[]{"-Xms768m", "-Xmx768m", "-XX:MaxDirectMemorySize=768m"};
+        } else {
+            customArgs = new String[]{"-Xms768m", "-Xmx768m", "-XX:MaxDirectMemorySize=768m", "-Djmh.executor=CUSTOM",
+                    "-Djmh.executor.class=io.netty.microbench.util.AbstractMicrobenchmark$HarnessExecutor"};
+        }
+        String[] jvmArgs = new String[BASE_JVM_ARGS.length + customArgs.length];
+        System.arraycopy(BASE_JVM_ARGS, 0, jvmArgs, 0, BASE_JVM_ARGS.length);
+        System.arraycopy(customArgs, 0, jvmArgs, BASE_JVM_ARGS.length, customArgs.length);
+        if (disableAssertions) {
+            jvmArgs = removeAssertions(jvmArgs);
+        }
+        this.jvmArgs = jvmArgs;
     }
 
     @Override
     protected String[] jvmArgs() {
-        if (!disableAssertions) {
-            return JVM_ARGS;
-        }
-
-        if (jvmArgsWithNoAssertions == null) {
-            jvmArgsWithNoAssertions = removeAssertions(JVM_ARGS);
-        }
-        return jvmArgsWithNoAssertions;
+        return jvmArgs;
     }
 
     @Override