@windie 2016-03-22
Before https://github.com/netty/netty/pull/5012/commits/5dcc0c9d05b9c1a421619f0f4a5fb19e85f06ddc
```
Benchmark                                            Mode  Cnt      Score     Error  Units
FastThreadLocalFastPathBenchmark.fastThreadLocal    thrpt   20  63891.724 ± 388.336  ops/s
FastThreadLocalFastPathBenchmark.jdkThreadLocalGet  thrpt   20  46496.464 ± 796.301  ops/s

Benchmark                                            Mode  Cnt      Score     Error  Units
FastThreadLocalSlowPathBenchmark.fastThreadLocal    thrpt   20  22163.357 ± 261.018  ops/s
FastThreadLocalSlowPathBenchmark.jdkThreadLocalGet  thrpt   20  46401.424 ± 780.948  ops/s
```
After
```
Benchmark                                            Mode  Cnt      Score      Error  Units
FastThreadLocalFastPathBenchmark.fastThreadLocal    thrpt   20  63397.962 ± 1205.249  ops/s
FastThreadLocalFastPathBenchmark.jdkThreadLocalGet  thrpt   20  46614.613 ±  776.756  ops/s

Benchmark                                            Mode  Cnt      Score     Error  Units
FastThreadLocalSlowPathBenchmark.fastThreadLocal    thrpt   20  30672.368 ± 269.152  ops/s
FastThreadLocalSlowPathBenchmark.jdkThreadLocalGet  thrpt   20  45984.833 ± 223.437  ops/s
```