@normanmaurer 2016-03-22
@windie I don't see any improvements in your posted benchmark output... What I'm missing?