@windie 2016-03-22
@normanmaurer Sorry that it's not clear. I copied the two benchmark outputs about the improvement here:

Before:
```
FastThreadLocalSlowPathBenchmark.fastThreadLocal    thrpt   20  22163.357 ± 261.018  ops/s
```
After:
```
FastThreadLocalSlowPathBenchmark.fastThreadLocal    thrpt   20  30672.368 ± 269.152  ops/s
```