@normanmaurer 2016-03-23
Cherry-picked into 4.1 (3ad55eb83915259dc9faef1a379ccd39fb35ec6f)  and 4.0 (f5b4937543455653341f3caea63400fb7b80cc99)